# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "default": "1",
            "title": "Api Version",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "JobCancelRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "default": "",
            "maxLength": 200,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "JobCancelRequest",
        "type": "object"
      },
      "JobData": {
        "additionalProperties": false,
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobPublicData"
          }
        },
        "required": [
          "job"
        ],
        "title": "JobData",
        "type": "object"
      },
      "JobEnvelope": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "default": "1",
            "title": "Api Version",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/JobData"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "data"
        ],
        "title": "JobEnvelope",
        "type": "object"
      },
      "JobErrorData": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "JobErrorData",
        "type": "object"
      },
      "JobPublicData": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "minimum": 0,
            "title": "Attempt",
            "type": "integer"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobErrorData"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "finished_at": {
            "title": "Finished At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "max_attempts": {
            "minimum": 1,
            "title": "Max Attempts",
            "type": "integer"
          },
          "progress": {
            "additionalProperties": true,
            "title": "Progress",
            "type": "object"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Result"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "cancel_requested",
          "attempt",
          "max_attempts",
          "created_at",
          "updated_at",
          "started_at",
          "finished_at",
          "expires_at"
        ],
        "title": "JobPublicData",
        "type": "object"
      },
      "JobResultData": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          }
        },
        "required": [
          "job_id",
          "result"
        ],
        "title": "JobResultData",
        "type": "object"
      },
      "JobResultEnvelope": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "default": "1",
            "title": "Api Version",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/JobResultData"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "data"
        ],
        "title": "JobResultEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "BLD API Key",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Stable API contract for internal tools and AI applications.",
    "title": "BLD Matcher API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnvelope"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read a persistent job owned by the API principal",
        "x-required-scopes": [
          "jobs:read"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnvelope"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Request cancellation of a persistent job",
        "x-required-scopes": [
          "jobs:cancel"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "operationId": "getJobResult",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultEnvelope"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read the completed result of a persistent job",
        "x-required-scopes": [
          "jobs:read"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}